backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# Redis
redis[hiredis]==5.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0

# Logging & Monitoring
structlog==24.4.0

# Rate Limiting
slowapi==0.1.9

# SSE
sse-starlette==2.2.1
